Ruby class for communicating with adc16_test model
Some of the functionality is generic to any ADC16 model.  Some is
specific to the adc16_test model (and should be moved to a subclass).

All of the cirrent implementation is based on the initial ADC16 yellow
block from gateware Jama and Genn.  It will need to be updated as the
ADC16 yellow block gateware evolves.