Avoid race condition when writing delay taps
When writing the delay tap values, the rising edge of the "reset" bit
triggers the current value to be read in.  Because the bits come from
the EPB clock domain, it's possible for the reset bit to be captured
into the fabric clock domain before all the delay value bits are
captured into the fabric clock domain.  The built-in edge detection
logic likely delays the triggereing long enough for this not to be an
issue, but writing the delay value first with all reset bits set to 0, then
writing the same delay value with the appropriate reset bits set to 1
avoids the race condition altogether.