Remove lane parameter from walk_taps
Since we can't distinguish which lane corresponds to even/odd samples,
there's no point in letting the caller think we can.  walk_taps now
returns a two element array.  The first element is an array of delay tap
values where there were no errors (i.e. unexpected values) in the 64 K
samples.  The second elements is a 32 element array of error counts.
Each element of the error count array is a two element array containing
the error counts for the even and odd samples for the given channel.

The return value is:

[
  good_taps,
  [
    [dt0_even_errcnt, dt0_odd_errcnt],
    [dt1_even_errcnt, dt1_odd_errcnt],
    ...,
    [dt31_even_errcnt, dt31_odd_errcnt]
  ]
]