Initial support for separate "a" and "b" delays
This modifies delay_tap to be compatible with newer gateware (as of
adc16_test_2013_Jan_19_0934) that supports separate "a" and "b" delays.
This initial support, however, still has delay_tap setting the "a" and
"b" delays to the same value.  It remains compatible with older
gateware, but delay_tap may change in the (near) future in a way that is
incompatible with older gateware.