Make ADC16#walk_taps more robust
If walk_taps finds that tap 0 or 31 results in the expected data
pattern, then the full eye pattern is not guaranteed to be sampled since
some of the "good" zone will be outside the delay tap range.  This
change to walk_taps handles this case by first testing taps 0 and 31.
If either tap results in the expected data pattern, then the chip (all
channels, all lanes) is bitslipped so that the expected pattern will be
shifted by an odd number of bits (either 1 or 3 depending on how many
bitslips have already been performed on that chip) thereby moving its
"good" zone to be fully within the delay tap range.