Fix programming of FPGA in adc16_plot_taps
If a BOF file is given, the script needs to program the FPGA *before*
reading the number of ADCs supported.  This is required both to ensure
that the device is programmed (in the case of running adc16_plot_taps
with a BOF file on an unprogrammed FPGA) and to get the number of ADCs
supported by the given BOF file rather than the previously programmed
BOF file.