Make delay tap calibration more robust
The original calibration routine kept an Array of good tap settings and
did not properly handle the case where more than one good range of tap
settings is found.  In that case, it would "give up" and leave the delay
tap set at 31.

The new calibration routine keeps an Array of Ranges of good tap
settings.  Then is finds the longest Range and sets the delay tap to the
midpoint of that range.  This is much more robust.