Make adc16_plot_taps.rb sync the chips
adc16_plot_taps.rb used to walk the delay taps only.  This left the
IODELAY blocks setup to deskew the data properly, but they were not
necessarily bitslipped to ensure proper framing the the 8-bit samples.
Now that ADC16#sync_chips is its own method, adc16_plot_taps.rb can use
it to bitslip the channels to ensure that the samples are being captured
correctly.